refactor: modularize parse.ts into separate modules
Split the monolithic parse.ts (1560 lines) into focused modules:
- parseHelpers.ts (255 lines): Low-level utilities (isTagNode, getStringAttr, etc.)
- parseSentinels.ts (125 lines): Sentinel value parsing (|SKIP|, |ABORT|)
- parseFields.ts (851 lines): Field type parsers (parseStringField, etc.)
- parse.ts (461 lines): Form/group parsing and main parseForm() function

This improves maintainability as more field types are added in the future.
