+++ packages/markform/src/engine/parse.ts
@@ -7,61 +7,22 @@
 import Markdoc from '@markdoc/markdoc';
 import type { Node } from '@markdoc/markdoc';
 
-import { AGENT_ROLE, DEFAULT_PRIORITY } from '../settings.js';
 import type {
-  ApprovalMode,
-  CheckboxesField,
-  CheckboxesValue,
-  CheckboxMode,
-  CheckboxValue,
-  DateField,
-  DateValue,
   DocumentationBlock,
   DocumentationTag,
-  Field,
   FieldGroup,
   FieldResponse,
-  FieldValue,
   FormSchema,
   Id,
   IdIndexEntry,
-  MultiSelectField,
-  MultiSelectValue,
   Note,
-  NumberField,
-  NumberValue,
-  Option,
   ParsedForm,
-  SingleSelectField,
-  SingleSelectValue,
-  StringField,
-  StringListField,
-  StringListValue,
-  StringValue,
-  UrlField,
-  UrlListField,
-  UrlListValue,
-  UrlValue,
-  ValidatorRef,
-  YearField,
-  YearValue,
 } from './coreTypes.js';
+import { parseField } from './parseFields.js';
+import { getStringAttr, getValidateAttr, isTagNode, ParseError } from './parseHelpers.js';
 
-// =============================================================================
-// Error Types
-// =============================================================================
-
-/** Parse error with source location info */
-export class ParseError extends Error {
-  constructor(
-    message: string,
-    public readonly line?: number,
-    public readonly col?: number,
-  ) {
-    super(message);
-    this.name = 'ParseError';
-  }
-}
+// Re-export ParseError for backward compatibility
+export { ParseError } from './parseHelpers.js';
 
 // =============================================================================
 // Frontmatter Parsing
@@ -117,1071 +78,6 @@ function extractFrontmatter(content: string): FrontmatterResult {
 }
 
 // =============================================================================
-// Checkbox Marker Parsing
-// =============================================================================
-
-/** Map checkbox marker to state value */
-const CHECKBOX_MARKERS: Record<string, CheckboxValue> = {
-  '[ ]': 'todo',
-  '[x]': 'done',
-  '[X]': 'done',
-  '[/]': 'incomplete',
-  '[*]': 'active',
-  '[-]': 'na',
-  '[y]': 'yes',
-  '[Y]': 'yes',
-  '[n]': 'no',
-  '[N]': 'no',
-};
-
-// Regex to extract checkbox marker and label from text content
-// Text is like "[ ] Label" or "[x] Label"
-const OPTION_TEXT_PATTERN = /^(\[[^\]]\])\s*(.*?)\s*$/;
-
-interface ParsedMarkerText {
-  marker: string;
-  label: string;
-}
-
-/**
- * Parse option text to extract marker and label.
- * Text is like "[ ] Label" or "[x] Label".
- */
-function parseOptionText(text: string): ParsedMarkerText | null {
-  const match = OPTION_TEXT_PATTERN.exec(text);
-  if (!match) {
-    return null;
-  }
-
-  const marker = match[1] ?? '';
-  const label = (match[2] ?? '').trim();
-
-  return { marker, label };
-}
-
-// =============================================================================
-// Markdoc Tag Processing
-// =============================================================================
-
-/**
- * Check if a node is a tag node with specific name.
- * Works with raw AST nodes (not transformed Tags).
- */
-function isTagNode(node: Node, name?: string): boolean {
-  if (typeof node !== 'object' || node === null) {
-    return false;
-  }
-  if (node.type === 'tag' && node.tag) {
-    return name === undefined || node.tag === name;
-  }
-  return false;
-}
-
-/**
- * Get string attribute value or undefined.
- */
-function getStringAttr(node: Node, name: string): string | undefined {
-  const value: unknown = node.attributes?.[name];
-  return typeof value === 'string' ? value : undefined;
-}
-
-/**
- * Get number attribute value or undefined.
- */
-function getNumberAttr(node: Node, name: string): number | undefined {
-  const value: unknown = node.attributes?.[name];
-  return typeof value === 'number' ? value : undefined;
-}
-
-/**
- * Get boolean attribute value or undefined.
- */
-function getBooleanAttr(node: Node, name: string): boolean | undefined {
-  const value: unknown = node.attributes?.[name];
-  return typeof value === 'boolean' ? value : undefined;
-}
-
-/**
- * Get validator references from validate attribute.
- * Handles both single string and array formats.
- */
-function getValidateAttr(node: Node): ValidatorRef[] | undefined {
-  const value: unknown = node.attributes?.validate;
-  if (value === undefined || value === null) {
-    return undefined;
-  }
-  if (Array.isArray(value)) {
-    return value as ValidatorRef[];
-  }
-  if (typeof value === 'string') {
-    return [value];
-  }
-  if (typeof value === 'object') {
-    // Single object validator ref like { id: "foo", param: 1 }
-    return [value as ValidatorRef];
-  }
-  return undefined;
-}
-
-/**
- * Parsed option item from AST.
- */
-interface ParsedOptionItem {
-  /** ID from {% #id %} annotation */
-  id: string | null;
-  /** Text content (e.g., "[ ] Label" or "[x] Label") */
-  text: string;
-}
-
-/**
- * Extract option items from node children (for option lists).
- * Works with raw AST nodes. Collects text and ID from list items.
- */
-function extractOptionItems(node: Node): ParsedOptionItem[] {
-  const items: ParsedOptionItem[] = [];
-
-  /**
-   * Collect all text content from a node tree into a single string.
-   */
-  function collectText(n: Node): string {
-    let text = '';
-
-    // Text nodes have content in attributes
-    if (n.type === 'text' && typeof n.attributes?.content === 'string') {
-      text += n.attributes.content;
-    }
-
-    // Softbreak is a newline
-    if (n.type === 'softbreak') {
-      text += '\n';
-    }
-
-    // Recurse into children
-    if (n.children && Array.isArray(n.children)) {
-      for (const c of n.children) {
-        text += collectText(c);
-      }
-    }
-
-    return text;
-  }
-
-  /**
-   * Traverse to find list items and extract their content.
-   */
-  function traverse(child: Node): void {
-    if (!child || typeof child !== 'object') {
-      return;
-    }
-
-    // List items contain the option text and ID
-    if (child.type === 'item') {
-      const text = collectText(child);
-      // Markdoc parses {% #id %} as an id attribute on the item
-      const id = typeof child.attributes?.id === 'string' ? child.attributes.id : null;
-      if (text.trim()) {
-        items.push({ id, text: text.trim() });
-      }
-      return; // Don't recurse further into item children
-    }
-
-    // Recurse into children
-    if (child.children && Array.isArray(child.children)) {
-      for (const c of child.children) {
-        traverse(c);
-      }
-    }
-  }
-
-  if (node.children && Array.isArray(node.children)) {
-    for (const child of node.children) {
-      traverse(child);
-    }
-  }
-
-  return items;
-}
-
-/**
- * Extract fence value from node children.
- * Looks for ```value code blocks.
- */
-function extractFenceValue(node: Node): string | null {
-  function traverse(child: Node): string | null {
-    if (!child || typeof child !== 'object') {
-      return null;
-    }
-
-    // Check if this is a fence node with language="value"
-    if (child.type === 'fence') {
-      const lang = child.attributes?.language as string | undefined;
-      if (lang === 'value') {
-        return typeof child.attributes?.content === 'string' ? child.attributes.content : null;
-      }
-    }
-
-    // Traverse children
-    if (child.children && Array.isArray(child.children)) {
-      for (const c of child.children) {
-        const result = traverse(c);
-        if (result !== null) {
-          return result;
-        }
-      }
-    }
-
-    return null;
-  }
-
-  if (node.children && Array.isArray(node.children)) {
-    for (const child of node.children) {
-      const result = traverse(child);
-      if (result !== null) {
-        return result;
-      }
-    }
-  }
-
-  return null;
-}
-
-// =============================================================================
-// Field Parsing
-// =============================================================================
-
-/** Sentinel values for text fields */
-const SENTINEL_SKIP = '|SKIP|';
-const SENTINEL_ABORT = '|ABORT|';
-
-/**
- * Parse a sentinel value with optional parenthesized reason.
- * Formats: `|SKIP|`, `|SKIP| (reason text)`, `|ABORT|`, `|ABORT| (reason text)`
- * Returns null if the content is not a sentinel.
- */
-function parseSentinel(content: string | null): { type: 'skip' | 'abort'; reason?: string } | null {
-  if (!content) {
-    return null;
-  }
-
-  const trimmed = content.trim();
-  const reasonPattern = /^\((.+)\)$/s;
-
-  // Check for |SKIP| with optional reason
-  if (trimmed.startsWith(SENTINEL_SKIP)) {
-    const rest = trimmed.slice(SENTINEL_SKIP.length).trim();
-    if (rest === '') {
-      return { type: 'skip' };
-    }
-    // Extract reason from parentheses
-    const match = reasonPattern.exec(rest);
-    if (match?.[1]) {
-      return { type: 'skip', reason: match[1].trim() };
-    }
-    // Invalid format - just |SKIP| with non-parenthesized content
-    return null;
-  }
-
-  // Check for |ABORT| with optional reason
-  if (trimmed.startsWith(SENTINEL_ABORT)) {
-    const rest = trimmed.slice(SENTINEL_ABORT.length).trim();
-    if (rest === '') {
-      return { type: 'abort' };
-    }
-    // Extract reason from parentheses
-    const match = reasonPattern.exec(rest);
-    if (match?.[1]) {
-      return { type: 'abort', reason: match[1].trim() };
-    }
-    // Invalid format - just |ABORT| with non-parenthesized content
-    return null;
-  }
-
-  return null;
-}
-
-/**
- * Check for sentinel values in fence content and validate against state attribute.
- * Handles the common pattern of checking for |SKIP| and |ABORT| sentinels in field values.
- *
- * @param node - The field node to check
- * @param fieldId - The field ID for error messages
- * @param required - Whether the field is required (skip not allowed on required fields)
- * @returns A FieldResponse if a sentinel is found, null otherwise
- */
-function tryParseSentinelResponse(
-  node: Node,
-  fieldId: string,
-  required: boolean,
-): FieldResponse | null {
-  const fenceContent = extractFenceValue(node);
-  const stateAttr = getStringAttr(node, 'state');
-
-  const sentinel = parseSentinel(fenceContent);
-  if (!sentinel) {
-    return null;
-  }
-
-  if (sentinel.type === 'skip') {
-    if (stateAttr !== undefined && stateAttr !== 'skipped') {
-      throw new ParseError(
-        `Field '${fieldId}' has conflicting state='${stateAttr}' with |SKIP| sentinel`,
-      );
-    }
-    if (required) {
-      throw new ParseError(
-        `Field '${fieldId}' is required but has |SKIP| sentinel. Cannot skip required fields.`,
-      );
-    }
-    return { state: 'skipped', ...(sentinel.reason && { reason: sentinel.reason }) };
-  }
-
-  if (sentinel.type === 'abort') {
-    if (stateAttr !== undefined && stateAttr !== 'aborted') {
-      throw new ParseError(
-        `Field '${fieldId}' has conflicting state='${stateAttr}' with |ABORT| sentinel`,
-      );
-    }
-    return { state: 'aborted', ...(sentinel.reason && { reason: sentinel.reason }) };
-  }
-
-  return null;
-}
-
-/**
- * Determine if a field value is empty.
- * For old forms without state attributes, this infers whether the field has been filled.
- */
-function isValueEmpty(value: FieldValue): boolean {
-  switch (value.kind) {
-    case 'string':
-    case 'number':
-    case 'url':
-    case 'date':
-    case 'year':
-      return value.value === null;
-    case 'string_list':
-    case 'url_list':
-      return value.items.length === 0;
-    case 'single_select':
-      return value.selected === null;
-    case 'multi_select':
-      return value.selected.length === 0;
-    case 'checkboxes': {
-      // Empty if all checkboxes are in default unchecked state (todo/unfilled)
-      const values = Object.values(value.values);
-      return values.every((v) => v === 'todo' || v === 'unfilled');
-    }
-  }
-}
-
-/**
- * Create a FieldResponse from a FieldValue.
- * For old forms without state attributes, infers state from value content.
- */
-function createFieldResponse(value: FieldValue): FieldResponse {
-  if (isValueEmpty(value)) {
-    return { state: 'unanswered' };
-  }
-  return { state: 'answered', value };
-}
-
-/**
- * Parse state attribute from field node and validate consistency.
- * Returns a FieldResponse based on state attribute, sentinel values, or value content.
- */
-function parseFieldResponse(
-  node: Node,
-  value: FieldValue,
-  fieldId: string,
-  required: boolean,
-): FieldResponse {
-  const stateAttr = getStringAttr(node, 'state');
-  const isFilled = !isValueEmpty(value);
-
-  // Validate state attribute value if present
-  if (stateAttr !== undefined) {
-    if (
-      stateAttr !== 'empty' &&
-      stateAttr !== 'answered' &&
-      stateAttr !== 'skipped' &&
-      stateAttr !== 'aborted'
-    ) {
-      throw new ParseError(
-        `Invalid state attribute '${stateAttr}' on field '${fieldId}'. Must be empty, answered, skipped, or aborted`,
-      );
-    }
-
-    // Validate state vs filled consistency
-    if (stateAttr === 'skipped' || stateAttr === 'aborted') {
-      if (isFilled) {
-        throw new ParseError(
-          `Field '${fieldId}' has state='${stateAttr}' but contains a value. ${stateAttr} fields cannot have values.`,
-        );
-      }
-    }
-
-    // Validate skipped on required fields
-    if (stateAttr === 'skipped' && required) {
-      throw new ParseError(
-        `Field '${fieldId}' is required but has state='skipped'. Cannot skip required fields.`,
-      );
-    }
-
-    // Return response based on explicit state attribute
-    if (stateAttr === 'skipped') {
-      return { state: 'skipped' };
-    }
-    if (stateAttr === 'aborted') {
-      return { state: 'aborted' };
-    }
-    if (stateAttr === 'empty') {
-      return { state: 'unanswered' };
-    }
-    if (stateAttr === 'answered') {
-      if (!isFilled) {
-        throw new ParseError(`Field '${fieldId}' has state='answered' but has no value`);
-      }
-      return { state: 'answered', value };
-    }
-  }
-
-  // No state attribute - infer from value content (backward compatibility)
-  return createFieldResponse(value);
-}
-
-/**
- * Get priority attribute value or default to DEFAULT_PRIORITY.
- */
-function getPriorityAttr(node: Node): 'high' | 'medium' | 'low' {
-  const value = getStringAttr(node, 'priority');
-  if (value === 'high' || value === 'medium' || value === 'low') {
-    return value;
-  }
-  return DEFAULT_PRIORITY;
-}
-
-/**
- * Parse a string-field tag.
- */
-function parseStringField(node: Node): { field: StringField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("string-field missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`string-field '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-
-  const field: StringField = {
-    kind: 'string',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    multiline: getBooleanAttr(node, 'multiline'),
-    pattern: getStringAttr(node, 'pattern'),
-    minLength: getNumberAttr(node, 'minLength'),
-    maxLength: getNumberAttr(node, 'maxLength'),
-    validate: getValidateAttr(node),
-  };
-
-  // Check for sentinel values first
-  const sentinelResponse = tryParseSentinelResponse(node, id, required);
-  if (sentinelResponse) {
-    return { field, response: sentinelResponse };
-  }
-
-  // No sentinel - parse normally
-  const fenceContent = extractFenceValue(node);
-  const trimmedContent = fenceContent !== null ? fenceContent.trim() : null;
-  const value: StringValue = {
-    kind: 'string',
-    value: trimmedContent,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a number-field tag.
- */
-function parseNumberField(node: Node): { field: NumberField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("number-field missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`number-field '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-
-  const field: NumberField = {
-    kind: 'number',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    min: getNumberAttr(node, 'min'),
-    max: getNumberAttr(node, 'max'),
-    integer: getBooleanAttr(node, 'integer'),
-    validate: getValidateAttr(node),
-  };
-
-  // Check for sentinel values first
-  const sentinelResponse = tryParseSentinelResponse(node, id, required);
-  if (sentinelResponse) {
-    return { field, response: sentinelResponse };
-  }
-
-  // No sentinel - parse number normally
-  const fenceContent = extractFenceValue(node);
-  const trimmedContent = fenceContent !== null ? fenceContent.trim() : '';
-  let numValue: number | null = null;
-  if (trimmedContent) {
-    const parsed = Number(trimmedContent);
-    if (!Number.isNaN(parsed)) {
-      numValue = parsed;
-    }
-  }
-
-  const value: NumberValue = {
-    kind: 'number',
-    value: numValue,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a string-list tag.
- */
-function parseStringListField(node: Node): { field: StringListField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("string-list missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`string-list '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-
-  const field: StringListField = {
-    kind: 'string_list',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    minItems: getNumberAttr(node, 'minItems'),
-    maxItems: getNumberAttr(node, 'maxItems'),
-    itemMinLength: getNumberAttr(node, 'itemMinLength'),
-    itemMaxLength: getNumberAttr(node, 'itemMaxLength'),
-    uniqueItems: getBooleanAttr(node, 'uniqueItems'),
-    validate: getValidateAttr(node),
-  };
-
-  // Check for sentinel values first
-  const sentinelResponse = tryParseSentinelResponse(node, id, required);
-  if (sentinelResponse) {
-    return { field, response: sentinelResponse };
-  }
-
-  // No sentinel - parse list normally
-  const fenceContent = extractFenceValue(node);
-  const items: string[] = [];
-  if (fenceContent !== null) {
-    const lines = fenceContent.split('\n');
-    for (const line of lines) {
-      const trimmed = line.trim();
-      if (trimmed) {
-        items.push(trimmed);
-      }
-    }
-  }
-
-  const value: StringListValue = {
-    kind: 'string_list',
-    items,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse options from a select/checkbox field.
- */
-function parseOptions(
-  node: Node,
-  fieldId: string,
-): { options: Option[]; selected: Record<string, CheckboxValue> } {
-  const items = extractOptionItems(node);
-  const options: Option[] = [];
-  const selected: Record<string, CheckboxValue> = {};
-  const seenIds = new Set<string>();
-
-  for (const item of items) {
-    const parsed = parseOptionText(item.text);
-    if (!parsed) {
-      continue;
-    }
-
-    if (!item.id) {
-      throw new ParseError(
-        `Option in field '${fieldId}' missing ID annotation. Use {% #option_id %}`,
-      );
-    }
-
-    if (seenIds.has(item.id)) {
-      throw new ParseError(`Duplicate option ID '${item.id}' in field '${fieldId}'`);
-    }
-    seenIds.add(item.id);
-
-    options.push({ id: item.id, label: parsed.label });
-
-    const state = CHECKBOX_MARKERS[parsed.marker];
-    if (state !== undefined) {
-      selected[item.id] = state;
-    }
-  }
-
-  return { options, selected };
-}
-
-/**
- * Parse a single-select tag.
- */
-function parseSingleSelectField(node: Node): { field: SingleSelectField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("single-select missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`single-select '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-  const { options, selected } = parseOptions(node, id);
-
-  const field: SingleSelectField = {
-    kind: 'single_select',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    options,
-    validate: getValidateAttr(node),
-  };
-
-  // Find the selected option (exactly one with done/[x] state)
-  let selectedOption: string | null = null;
-  for (const [optId, state] of Object.entries(selected)) {
-    if (state === 'done') {
-      selectedOption = optId;
-      break;
-    }
-  }
-
-  const value: SingleSelectValue = {
-    kind: 'single_select',
-    selected: selectedOption,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a multi-select tag.
- */
-function parseMultiSelectField(node: Node): { field: MultiSelectField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("multi-select missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`multi-select '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-  const { options, selected } = parseOptions(node, id);
-
-  const field: MultiSelectField = {
-    kind: 'multi_select',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    options,
-    minSelections: getNumberAttr(node, 'minSelections'),
-    maxSelections: getNumberAttr(node, 'maxSelections'),
-    validate: getValidateAttr(node),
-  };
-
-  // Collect all selected options (those with done/[x] state)
-  const selectedOptions: string[] = [];
-  for (const [optId, state] of Object.entries(selected)) {
-    if (state === 'done') {
-      selectedOptions.push(optId);
-    }
-  }
-
-  const value: MultiSelectValue = {
-    kind: 'multi_select',
-    selected: selectedOptions,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a checkboxes tag.
- */
-function parseCheckboxesField(node: Node): { field: CheckboxesField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("checkboxes missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`checkboxes '${id}' missing required 'label' attribute`);
-  }
-
-  const { options, selected } = parseOptions(node, id);
-
-  const checkboxModeStr = getStringAttr(node, 'checkboxMode');
-  let checkboxMode: CheckboxMode = 'multi'; // default
-  if (
-    checkboxModeStr === 'multi' ||
-    checkboxModeStr === 'simple' ||
-    checkboxModeStr === 'explicit'
-  ) {
-    checkboxMode = checkboxModeStr;
-  }
-
-  const approvalModeStr = getStringAttr(node, 'approvalMode');
-  let approvalMode: ApprovalMode = 'none'; // default
-  if (approvalModeStr === 'blocking') {
-    approvalMode = 'blocking';
-  }
-
-  // Handle required attribute based on checkboxMode:
-  // - explicit mode is inherently required (cannot be set to false)
-  // - multi/simple modes default to optional (false)
-  const explicitRequired = getBooleanAttr(node, 'required');
-  let required: boolean;
-  if (checkboxMode === 'explicit') {
-    if (explicitRequired === false) {
-      throw new ParseError(
-        `Checkbox field "${label}" has checkboxMode="explicit" which is inherently required. ` +
-          `Cannot set required=false. Remove required attribute or change checkboxMode.`,
-      );
-    }
-    required = true; // explicit mode is always required
-  } else {
-    required = explicitRequired ?? false; // multi/simple default to optional
-  }
-
-  const field: CheckboxesField = {
-    kind: 'checkboxes',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    checkboxMode,
-    minDone: getNumberAttr(node, 'minDone'),
-    options,
-    approvalMode,
-    validate: getValidateAttr(node),
-  };
-
-  // Initialize all options to their default state based on mode
-  const values: Record<string, CheckboxValue> = {};
-
-  for (const opt of options) {
-    const state = selected[opt.id];
-    if (state === undefined || state === 'todo') {
-      // For explicit mode, "todo" (from [ ]) means "unfilled"
-      values[opt.id] = checkboxMode === 'explicit' ? 'unfilled' : 'todo';
-    } else {
-      values[opt.id] = state;
-    }
-  }
-
-  const value: CheckboxesValue = {
-    kind: 'checkboxes',
-    values,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a url-field tag.
- */
-function parseUrlField(node: Node): { field: UrlField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("url-field missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`url-field '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-
-  const field: UrlField = {
-    kind: 'url',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    validate: getValidateAttr(node),
-  };
-
-  // Check for sentinel values first
-  const sentinelResponse = tryParseSentinelResponse(node, id, required);
-  if (sentinelResponse) {
-    return { field, response: sentinelResponse };
-  }
-
-  // No sentinel - parse normally
-  const fenceContent = extractFenceValue(node);
-  const trimmedContent = fenceContent !== null ? fenceContent.trim() : null;
-  const value: UrlValue = {
-    kind: 'url',
-    value: trimmedContent,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a url-list tag.
- */
-function parseUrlListField(node: Node): { field: UrlListField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("url-list missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`url-list '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-
-  const field: UrlListField = {
-    kind: 'url_list',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    minItems: getNumberAttr(node, 'minItems'),
-    maxItems: getNumberAttr(node, 'maxItems'),
-    uniqueItems: getBooleanAttr(node, 'uniqueItems'),
-    validate: getValidateAttr(node),
-  };
-
-  // Check for sentinel values first
-  const sentinelResponse = tryParseSentinelResponse(node, id, required);
-  if (sentinelResponse) {
-    return { field, response: sentinelResponse };
-  }
-
-  // No sentinel - parse list normally
-  const fenceContent = extractFenceValue(node);
-  const items: string[] = [];
-  if (fenceContent !== null) {
-    const lines = fenceContent.split('\n');
-    for (const line of lines) {
-      const trimmed = line.trim();
-      if (trimmed) {
-        items.push(trimmed);
-      }
-    }
-  }
-
-  const value: UrlListValue = {
-    kind: 'url_list',
-    items,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a date-field tag.
- */
-function parseDateField(node: Node): { field: DateField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("date-field missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`date-field '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-
-  const field: DateField = {
-    kind: 'date',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    min: getStringAttr(node, 'min'),
-    max: getStringAttr(node, 'max'),
-    validate: getValidateAttr(node),
-  };
-
-  // Check for sentinel values first
-  const sentinelResponse = tryParseSentinelResponse(node, id, required);
-  if (sentinelResponse) {
-    return { field, response: sentinelResponse };
-  }
-
-  // No sentinel - parse normally
-  // Convert empty strings to null explicitly (nullish coalescing doesn't work for empty strings)
-  const fenceContent = extractFenceValue(node);
-  let dateValue: string | null = null;
-  if (fenceContent !== null) {
-    const trimmed = fenceContent.trim();
-    if (trimmed.length > 0) {
-      dateValue = trimmed;
-    }
-  }
-  const value: DateValue = {
-    kind: 'date',
-    value: dateValue,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a year-field tag.
- */
-function parseYearField(node: Node): { field: YearField; response: FieldResponse } {
-  const id = getStringAttr(node, 'id');
-  const label = getStringAttr(node, 'label');
-
-  if (!id) {
-    throw new ParseError("year-field missing required 'id' attribute");
-  }
-  if (!label) {
-    throw new ParseError(`year-field '${id}' missing required 'label' attribute`);
-  }
-
-  const required = getBooleanAttr(node, 'required') ?? false;
-
-  const field: YearField = {
-    kind: 'year',
-    id,
-    label,
-    required,
-    priority: getPriorityAttr(node),
-    role: getStringAttr(node, 'role') ?? AGENT_ROLE,
-    min: getNumberAttr(node, 'min'),
-    max: getNumberAttr(node, 'max'),
-    validate: getValidateAttr(node),
-  };
-
-  // Check for sentinel values first
-  const sentinelResponse = tryParseSentinelResponse(node, id, required);
-  if (sentinelResponse) {
-    return { field, response: sentinelResponse };
-  }
-
-  // No sentinel - parse year normally
-  const fenceContent = extractFenceValue(node);
-  const trimmedContent = fenceContent !== null ? fenceContent.trim() : '';
-  let yearValue: number | null = null;
-  if (trimmedContent) {
-    const parsed = Number.parseInt(trimmedContent, 10);
-    if (!Number.isNaN(parsed)) {
-      yearValue = parsed;
-    }
-  }
-
-  const value: YearValue = {
-    kind: 'year',
-    value: yearValue,
-  };
-
-  const response = parseFieldResponse(node, value, id, required);
-  return { field, response };
-}
-
-/**
- * Parse a field tag and return field schema and response.
- */
-function parseField(node: Node): { field: Field; response: FieldResponse } | null {
-  if (!isTagNode(node)) {
-    return null;
-  }
-  switch (node.tag) {
-    case 'string-field':
-      return parseStringField(node);
-    case 'number-field':
-      return parseNumberField(node);
-    case 'string-list':
-      return parseStringListField(node);
-    case 'single-select':
-      return parseSingleSelectField(node);
-    case 'multi-select':
-      return parseMultiSelectField(node);
-    case 'checkboxes':
-      return parseCheckboxesField(node);
-    case 'url-field':
-      return parseUrlField(node);
-    case 'url-list':
-      return parseUrlListField(node);
-    case 'date-field':
-      return parseDateField(node);
-    case 'year-field':
-      return parseYearField(node);
-    default:
-      return null;
-  }
-}
-
-// =============================================================================
 // Group and Form Parsing
 // =============================================================================
 
@@ -1216,7 +112,7 @@ function parseFieldGroup(
 
   idIndex.set(id, { nodeType: 'group', parentId });
 
-  const children: Field[] = [];
+  const children: FieldGroup['children'] = [];
 
   // Traverse children to find fields
   function processChildren(child: Node): void {
